[AndroidPageServer] Create Initial Android App
[Status]:
Not Working as Expected

[What Works]:
If 2 Devices are connected on same Wifi Network,
The Static HTML can be viewed by another device.

[What Needs Fixing]:
It needs to work with Hotspot of the first Device
as well.

[Possible Fixes]:
Look into LocalOnlyHotspot, and how to implement it
Also check its scope

[Secondary Fixes Required]:
Currently using Nanohttpd. Need to move away from it.

Signed-off-by: ankit.roy <cackrogt93@gmail.com>