[AndroidPageServer] Make the app work with LocalHotspot
[Status]:
Working with local hotspot

[What Works]:
If 2 Devices are connected on 1st devices Hotspot
we get the shared webpage using hotspot based option.

[What Needs Fixing]:
1. Sometimes the Addr is in use, which was causing exception
(have added Log instead of Throw RuntimeException
2. When you press the old button for normal Server start,
the value of 0.0.0.0:8090 is given, while the hosting happens
elsewhere(192.168.123.11:8090 in current tests)
This issue is not there when you press the hotspot based button.

[Possible Fixes]:
1. Just check for other bugs or crashes in 1st case.
2. In this case, the fixed code method is already there,
check if we can just use it in the old case as well.
other wise we will have to differentiate between 2 cases.

[Secondary Fixes Required]:
Currently using Nanohttpd. Need to move away from it.
Location Permission is being taken.

Signed-off-by: ankit.roy <cackrogt93@gmail.com>